Update annotatorApp.groovy
Removed alert box which reports the new PathClass of the selected annotation (i.e. a named PathClass with intensity derivation) to make UI experience smoother. Post-dialog alert will now only be shown when the user supplies an invalid input combination (in particular, not selecting a PathClass or an intensity, or selecting the null PathClass), or when the user explicitly cancels out of the dialog.