Update tileApp.groovy
Streamlined the APIs for the dialogCallable, alertCallable, and the interaction with Params; created OptionBox and OptionBox.OptionCallable to override old boilerplate for using select params in dialogs. Added feature to allow user to choose from a) keep full pathClass names, b) keep only the base pathClass names, c) do not install any pathClasses for the annotations.